lsp: split analysis responsibilities
