+++ purple-garden-cli/src/lsp.rs
@@ -1,7 +1,10 @@
 mod analysis;
 mod code_action;
+mod collect;
 mod completion;
+mod definition;
 mod diagnostic;
+mod hover;
 mod server;
 mod source;
